+++ tests/testthat/test-join.R
@@ -1,7 +1,7 @@
 test_that("matrixset row info join works", {
 
   ms1 <- remove_row_annotation(student_results, class, teacher)
-  ms <- join_row_info(ms1, student_results)
+  suppressWarnings(ms <- join_row_info(ms1, student_results))
   ms_ref <- student_results
   ri <- row_info(ms_ref)
   ri <- dplyr::mutate(ri, previous_year_score.y = previous_year_score)
@@ -25,9 +25,159 @@ test_that("matrixset row info join works", {
                "variable \"previous_score\" is not a known trait")
 
 
+  meta <- tibble::tibble(sample = c("student 2", "student 2"),
+                         msr = c("height", "weight"),
+                         value = c(145, 32))
+  expect_error(join_row_info(student_results, meta, by = c(".rowname"="sample")),
+               "'by' does not result in unique row names")
+
+
+  expect_warning(join_row_info(student_results, meta, by = c(".rowname"="sample"),
+                               adjust = TRUE, names_glue = TRUE),
+                 "row names \\(\\.rowname\\) have changed following matrix adjustment")
+
+
+  suppressWarnings(ms <- join_row_info(student_results, meta, by = c(".rowname"="sample"),
+                      adjust = TRUE, names_glue = TRUE))
+  ms_ref_inf <- dplyr::left_join(row_info(student_results), meta,
+                                 by = c(".rowname"="sample"))
+  nms <- ms_ref_inf$.rowname
+  idx <- match(nms, rownames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], unq_id[unq_id > 0], sep = "..")
+  ms_ref_inf$.rowname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][idx, ]
+    rownames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = ms_ref_inf, row_key = ".rowname",
+                      column_info = column_info(student_results),
+                      column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+
+  suppressWarnings(ms <- join_row_info(student_results, meta, by = c(".rowname"="sample"),
+                      adjust = TRUE, names_glue = "{.tag}"))
+  expect_identical(ms, ms_ref)
+
+  suppressWarnings(ms <- join_row_info(student_results, meta, by = c(".rowname"="sample"),
+                      adjust = TRUE, names_glue = "{.tag}_{msr}"))
+  ms_ref_inf <- dplyr::left_join(row_info(student_results), meta,
+                                 by = c(".rowname"="sample"))
+  nms <- ms_ref_inf$.rowname
+  idx <- match(nms, rownames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], ms_ref_inf$msr[unq_id > 0], sep = "_")
+  ms_ref_inf$.rowname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][idx, ]
+    rownames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = ms_ref_inf, row_key = ".rowname",
+                      column_info = column_info(student_results),
+                      column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+  expect_error(purrr::quietly(join_row_info)(student_results, meta, by = c(".rowname"="sample"),
+                             adjust = TRUE, names_glue = "{.tag}_{msr2}"),
+               "object 'msr2' not found")
+
+
+  meta <- tibble::tibble(sample = c("student 2", "student 2"),
+                         class = c("classA", "classA"),
+                         msr = c("height", "weight"),
+                         value = c(145, 32))
+  suppressWarnings(ms <- join_row_info(student_results, meta, by = c(".rowname"="sample"),
+                      adjust = TRUE, names_glue = TRUE))
+  ms_ref_inf <- dplyr::left_join(row_info(student_results), meta,
+                                 by = c(".rowname"="sample"))
+  nms <- ms_ref_inf$.rowname
+  idx <- match(nms, rownames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], unq_id[unq_id > 0], sep = "..")
+  ms_ref_inf$.rowname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][idx, ]
+    rownames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = ms_ref_inf, row_key = ".rowname",
+                      column_info = column_info(student_results),
+                      column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+
+  suppressWarnings(ms <- join_row_info(student_results, meta, by = c(".rowname"="sample"),
+                      adjust = TRUE, names_glue = "{.tag}"))
+  expect_identical(ms, ms_ref)
+
+  w <- purrr::quietly(join_row_info)(student_results, meta,
+                                     by = c(".rowname"="sample", "class"),
+                                     adjust = TRUE, names_glue = TRUE)
+  w <- w$warnings
+  expect_true(any(w == "some traits have changed type ('class')"))
+
+
+  suppressWarnings(ms <- join_row_info(student_results, meta, by = c(".rowname"="sample", "class"),
+                      adjust = TRUE, names_glue = TRUE))
+  ms_ref_inf <- dplyr::left_join(row_info(student_results), meta,
+                                 by = c(".rowname"="sample", "class"))
+  nms <- ms_ref_inf$.rowname
+  idx <- match(nms, rownames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], unq_id[unq_id > 0], sep = "..")
+  ms_ref_inf$.rowname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][idx, ]
+    rownames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = ms_ref_inf, row_key = ".rowname",
+                      column_info = column_info(student_results),
+                      column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+
+  suppressWarnings(ms <- join_row_info(student_results, meta, by = c(".rowname"="sample"),
+                      adjust = TRUE, names_glue = "{.tag}_{msr}"))
+  ms_ref_inf <- dplyr::left_join(row_info(student_results), meta,
+                                 by = c(".rowname"="sample"))
+  nms <- ms_ref_inf$.rowname
+  idx <- match(nms, rownames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], ms_ref_inf$msr[unq_id > 0], sep = "_")
+  ms_ref_inf$.rowname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][idx, ]
+    rownames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = ms_ref_inf, row_key = ".rowname",
+                      column_info = column_info(student_results),
+                      column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+
+  w <- purrr::quietly(join_row_info)(row_group_by(student_results, class), meta,
+                by = c(".rowname"="sample"), adjust = TRUE, names_glue = TRUE)
+  w <- w$warnings
+  expect_true(any(w == "grouping structure of '.ms_x' has changed."))
+
 
   ms1 <- remove_column_annotation(student_results, national_average, school_average)
-  ms <- join_column_info(ms1, student_results)
+  suppressWarnings(ms <- join_column_info(ms1, student_results))
   ms_ref <- student_results
   ci <- column_info(ms_ref)
   ci <- dplyr::mutate(ci, program.y = program)
@@ -38,9 +188,126 @@ test_that("matrixset row info join works", {
 
 
 
+  meta <- tibble::tibble(field = c("Mathematics", "Mathematics"),
+                         level = c("regular", "advanced"),
+                         value = c(100, 250))
+  expect_error(join_column_info(student_results, meta, by = c(".colname"="field")),
+               "'by' does not result in unique col names")
+
+  expect_warning(join_column_info(student_results, meta, by = c(".colname"="field"),
+                                  adjust = TRUE, names_glue = TRUE),
+                 "col names \\(\\.colname\\) have changed following matrix adjustment")
+
+  suppressWarnings(ms <- join_column_info(student_results, meta, by = c(".colname"="field"),
+                         adjust = TRUE, names_glue = TRUE))
+  ms_ref_inf <- dplyr::left_join(column_info(student_results), meta,
+                                 by = c(".colname"="field"))
+  nms <- ms_ref_inf$.colname
+  idx <- match(nms, colnames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], unq_id[unq_id > 0], sep = "..")
+  ms_ref_inf$.colname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][, idx]
+    colnames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = row_info(student_results),
+                      row_key = ".rowname",
+                      column_info = ms_ref_inf,
+                      column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+
+  suppressWarnings(ms <- join_column_info(student_results, meta, by = c(".colname"="field"),
+                         adjust = TRUE, names_glue = "{.tag}"))
+  expect_identical(ms, ms_ref)
+
+  suppressWarnings(ms <- join_column_info(student_results, meta, by = c(".colname"="field"),
+                         adjust = TRUE, names_glue = "{.tag}_{level}"))
+  ms_ref_inf <- dplyr::left_join(column_info(student_results), meta,
+                                 by = c(".colname"="field"))
+  nms <- ms_ref_inf$.colname
+  idx <- match(nms, colnames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], ms_ref_inf$level[unq_id > 0], sep = "_")
+  ms_ref_inf$.colname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][, idx]
+    colnames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = row_info(student_results),
+                      row_key = ".rowname",
+                      column_info = ms_ref_inf, column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+  expect_error(purrr::quietly(join_column_info)(student_results, meta, by = c(".colname"="field"),
+                                adjust = TRUE, names_glue = "{.tag}_{level2}"),
+               "object 'level2' not found")
+
+
+  meta <- tibble::tibble(field = c("Mathematics", "Mathematics"),
+                         program = c("Applied Science", "Applied Science"),
+                         level = c("regular", "advanced"),
+                         value = c(100, 250))
+  suppressWarnings(ms <- join_column_info(student_results, meta, by = c(".colname"="field"),
+                                          adjust = TRUE, names_glue = TRUE))
+  ms_ref_inf <- dplyr::left_join(column_info(student_results), meta,
+                                 by = c(".colname"="field"))
+  nms <- ms_ref_inf$.colname
+  idx <- match(nms, colnames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], unq_id[unq_id > 0], sep = "..")
+  ms_ref_inf$.colname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][, idx]
+    colnames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = row_info(student_results),
+                      row_key = ".rowname", column_info = ms_ref_inf,
+                      column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+
+  suppressWarnings(ms <- join_column_info(student_results, meta, by = c(".colname"="field"),
+                                          adjust = TRUE, names_glue = "{.tag}"))
+  expect_identical(ms, ms_ref)
+
+  suppressWarnings(ms <- join_column_info(student_results, meta, by = c(".colname"="field"),
+                                          adjust = TRUE, names_glue = "{.tag}_{level}"))
+  ms_ref_inf <- dplyr::left_join(column_info(student_results), meta,
+                                 by = c(".colname"="field"))
+  nms <- ms_ref_inf$.colname
+  idx <- match(nms, colnames(student_results))
+  unq_id <- unique_id(nms, length(nms))
+  nms[unq_id > 0] <- paste(nms[unq_id > 0], ms_ref_inf$level[unq_id > 0], sep = "_")
+  ms_ref_inf$.colname <- nms
+  m <- student_results[,,, keep_annotation = FALSE, warn_class_change = FALSE]
+  m <- lapply(seq_along(m), function(mi) {
+    M <- m[[mi]][, idx]
+    colnames(M) <- nms
+    M
+  })
+  names(m) <- matrixnames(student_results)
+  ms_ref <- matrixset(m, row_info = row_info(student_results),
+                      row_key = ".rowname", column_info = ms_ref_inf,
+                      column_key = ".colname")
+  expect_identical(ms, ms_ref)
+
+
+
+
   ms1 <- remove_row_annotation(filter_row(student_results, class %in% c("classA", "classC")),
                                class, teacher)
-  ms <- join_row_info(ms1, student_results)
+  suppressWarnings(ms <- join_row_info(ms1, student_results))
   ms_ref <- filter_row(student_results, class %in% c("classA", "classC"))
   ri <- row_info(ms_ref)
   ri <- dplyr::mutate(ri, previous_year_score.y = previous_year_score)
@@ -74,7 +341,7 @@ test_that("matrixset row info join works", {
 
   ms1 <- remove_column_annotation(filter_column(student_results, program == "Applied Science"),
                                   program)
-  ms <- join_column_info(ms1, student_results)
+  suppressWarnings(ms <- join_column_info(ms1, student_results))
   ms_ref <- filter_column(student_results, program == "Applied Science")
   ci <- column_info(ms_ref)
   ci <- dplyr::mutate(ci, national_average.y = national_average,
@@ -186,7 +453,7 @@ expect_identical(ms, ms_ref)
 
 
   ms3 <- remove_row_annotation(student_results, class)
-  ms <- join_row_info(row_group_by(ms3, teacher), student_results, by = c(".rowname", "previous_year_score"))
+  suppressWarnings(ms <- join_row_info(row_group_by(ms3, teacher), student_results, by = c(".rowname", "previous_year_score")))
   ms_ref <- student_results
   ri <- row_info(ms_ref)
   ri$teacher.x <- ri$teacher